requirements.txt
boto3~=1.24.74
requests~=2.28.1
pyyaml